Add files via upload
Initial can bus decoder for Frame ID 209 which represents 4 different wheel heights as 6 bit signals starting at Bit 40.   SavvyCan was used to identify the four signals however I am very new at this and it may not actually be the best signal available.  It troubles me that these signals only vary by 10 bits (decimal) between the Low and Very High displacement positions.  This does not seem like very good resolution but the mid position is apparently 33 allowing +- 30 bits deflection up or down due to dynamic response.  This frame ID is broadcasted about every 50 mS ( 20 Hz) so it must be important.   Empirically the offset variation of each signal suggests a static calibration value exists somewhere else on the bus.  Until it is found it does not make sense to add a scaling factor but it appears to be 4mm/bit plus offset("mm", "tbd").

Bits 0 thru 39 are active and likely related but have not yet been decoded.